Refactor data sources
